Hook DiscoveryResult.

+++ presence/rust/engine/presence_java_ffi/java/PresenceDiscoveryResult.java
@@ -1,6 +1,6 @@
 package com.google.nearby.presence.engine;
 public class PresenceDiscoveryResult {
-  public static void build(long medium) {
+  public static void build(int medium) {
     System.out.println("PresenceDiscoveryResult build with medium: " + medium);
   }
 }
